.gitlab/scripts/collect_container_images/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-gitlab==2.7.1
requests==2.25.1
requests-toolbelt==0.9.1
urllib3==1.26.4
